refactor: :truck: rename files from `raw` to `batch`
